requirements.txt
# Vehicle Damage Detection Application - Requirements
# Complete Research Paper Implementation
# Python 3.10+

# Deep Learning Framework
torch==2.7.1
torchvision==0.22.1
pillow==12.0.0

# Segmentation Models
segmentation-models-pytorch==0.3.3
timm==0.9.12  # For EfficientNet and other encoders
pretrainedmodels==0.7.4

# YOLO for Object Detection
ultralytics==8.3.61
opencv-python==4.10.0.84

# Data Augmentation
albumentations==1.4.0

# Web Framework
Flask==3.1.2
flask-cors==6.0.1
Werkzeug==3.1.3

# Scientific Computing
numpy==2.2.6
scipy==1.15.2

# Visualization
matplotlib==3.10.1
seaborn==0.13.2

# Data Processing
pandas==2.2.3
scikit-learn==1.6.1

# Utilities
pyyaml==6.0.2
tqdm==4.67.1
Jinja2==3.1.6
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0

# Image Processing
opencv-python-headless==4.10.0.84  # Headless version for servers
scikit-image==0.24.0

# For confusion matrix and reporting
prettytable==3.12.0
